- added General statistics menu - Modified elpaso not to use roman tribes when loaded   as scenario
git-svn-id: https://widelands.svn.sourceforge.net/svnroot/widelands/trunk@766 37b2a8de-5219-0410-9f54-a31bc463ab9c

+++ src/worker.cc
@@ -3223,7 +3223,7 @@ void Worker::fugitive_update(Game* g, State* state)
 		{
 			bool use = false;
 
-			if ((g->logic_rand() % 30) <= (30 - bestdist))
+			if ( (int)((g->logic_rand() % 30)) <= (30 - bestdist))
 				use = true;
 
 			// okay, move towards the flag of this warehouse
